feat(logger):⚡add logger utility and update types (#8)
* refactor(types): :recycle: back to use enum instead

* feat(logger): add logger utility and update types

This commit adds a new logger utility to the codebase, which includes console logging and file logging functionalities. It also updates the types to include a new logger function in the VNPayConfig type. These changes enhance the logging capabilities of the application.

* fix(example): :ambulance: fix example data to use enum

* refactor(types): update BuildPaymentUrlOptions logger field

This commit updates the logger field in the BuildPaymentUrlOptions type to support three different options: AllFields, OmitFields, and PickFields. This change enhances the flexibility and configurability of the logger in the VNPay.

* fix(logger): :ambulance: fix logger pick & omit issue

* test(logger): :white_check_mark: add test for logging in `buildPaymentUrl`

* chore: update configuration for coderabbitai

The commit updates the .coderabbit.yaml configuration file to include new settings for the language server, early access, reviews, auto reply in chat, and path instructions for code review. This improves the configuration and functionality of the CodeRabbit integration.

* refactor(logger): improve error handling in fileLogger

The commit refactors the `fileLogger` function in `logger.ts` to improve error handling. It adds an optional `errorCallback` parameter to the function signature and checks if `err` exists before logging the error and throwing an exception. If `errorCallback` is provided and `err` exists, it calls the `errorCallback` function with the error as an argument. This change enhances the error handling capabilities of the logger.

* chore: update configuration for coderabbitai

* style: :lipstick: add `eslint` & `prettier` config

* refactor(enums): add VnpCardType enum

The commit adds a new enum called VnpCardType to the enums/index.ts file. This enum includes two card types: ATM and QRCODE. This change enhances the flexibility and configurability of the VNPay application.

* refactor(types): update ReturnQueryFromVNPay type

This commit updates the ReturnQueryFromVNPay type in return-from-vnpay.type.ts to include a new property vnp_Amount, which represents the payment amount. This change enhances the clarity and completeness of the type definition.

* refactor(types): update LoggerOptions logger field

This commit updates the logger field in the LoggerOptions type to support three different options: 'all', 'omit', and 'pick'. It also adds an optional custom logger function for each option. This change enhances the flexibility and configurability of the logger in the codebase.

* refactor(types): update BuildPaymentUrlOptions logger field

* refactor(logger): add jsdocs

* refactor(logger): add logger

* feat: add regex constant for number validation

* refactor(utils): add ignoreLogger to BuildPaymentUrl test

* test(ipn): :white_check_mark: add verifyIpnCall test

This commit adds a new test for the verifyIpnCall function in the verify-ipn-call.test.ts file. The test verifies that the function returns the correct data with the given input. This change enhances the test coverage and ensures the correctness of the verifyIpnCall function.

* refactor(types): update jsdocs

* chore(example): :zap: update example with logger features

* chore: remove unused npm dependency portfinder

* refactor: remove unused code and update JSDocs

* docs(docs): :memo: typedoc generated

* docs(docs): :memo: typedoc generated

* chore: Update BuildPaymentUrlOptions property name

The `hashInUrl` property in `BuildPaymentUrlOptions` has been renamed to `withHash` for clarity and consistency with other properties.

* docs(docs): :memo: update docs for log feature

* docs(docs): :memo: typedoc generated

* refactor(docs): :fire: remove typedoc generated

* docs(docs): :memo: update docs for log feature